+++ R/get_conferences.R
@@ -2,27 +2,19 @@ get_conferences <- function(my_xml) {
 
   all_conf <- xml2::xml_find_all(my_xml, ".//APRESENTACAO-DE-TRABALHO")
 
-  if (length(all_conf) != 0) {
+  if (length(all_conf) == 0) return(tibble::tibble())
 
-    df_conf <- tibble::tibble()
-
-    for (i_node in all_conf) {
-      conf_dados_basicos <- fetch_df(i_node, ".//DADOS-BASICOS-DA-APRESENTACAO-DE-TRABALHO")
-
-      conf_detalhes <- fetch_df(i_node, ".//DETALHAMENTO-DA-APRESENTACAO-DE-TRABALHO")
-
-      df_conf <- dplyr::bind_rows(
-        df_conf,
-        dplyr::bind_cols(
-          conf_dados_basicos,
-          conf_detalhes
+  df_conf <- dplyr::bind_rows(
+    lapply(all_conf, function(node) {
+      fetch_node_df(
+        node,
+        c(
+          ".//DADOS-BASICOS-DA-APRESENTACAO-DE-TRABALHO",
+          ".//DETALHAMENTO-DA-APRESENTACAO-DE-TRABALHO"
         )
       )
-    }
-
-  } else {
-    df_conf <- tibble::tibble()
-  }
+    })
+  )
 
   return(df_conf)
 }